Cleanup: move StatusToolTip and helpers files into the shared dir
